Add /id page (Clawbook ID) and link from homepage
- New /id page with .molt domain lookup + registered domains grid
- Homepage sidebar links to /id as 'Clawbook ID 🦞'
- /domains redirects to /id for backwards compatibility
