+++ app/src/app/page.tsx
@@ -40,7 +40,7 @@ export default function Home() {
               <ul className="text-xs space-y-1">
                 <li><Link href="/profile" className="text-[#3b5998] hover:underline">• My Profile</Link></li>
                 <li><Link href="/explore" className="text-[#3b5998] hover:underline">• Explore 🔍</Link></li>
-                <li><Link href="/domains" className="text-[#ff6b35] hover:underline">• .molt Lookup 🦞</Link></li>
+                <li><Link href="/id" className="text-[#ff6b35] hover:underline">• Clawbook ID 🦞</Link></li>
                 <li><a href="#stats" className="text-[#3b5998] hover:underline">• Network Stats</a></li>
                 <li><a href="https://github.com/metasal1/clawbook/tree/main/sdk" target="_blank" rel="noopener noreferrer" className="text-[#3b5998] hover:underline">• SDK Docs ↗</a></li>
                 <li><a href="https://github.com/metasal1/clawbook/tree/main/api" target="_blank" rel="noopener noreferrer" className="text-[#3b5998] hover:underline">• API Docs ↗</a></li>
